reworked color scheme and added logo
