+++ src/styles/partials/_global.scss
@@ -17,7 +17,7 @@
     font-size: 1.6rem;
     margin: 0;
     background: $primary;
-    color: $secondary;
+    color: $quad;
     // padding: $mobile-pad;
     @include mobile-height
