+++ src/styles/partials/_variables.scss
@@ -1,14 +1,12 @@
 $mobile-min: 320px;
 $mobile-max: 767px;
 
-// $primary: #8caa87;
-$primary: #F6F1F1;
-// $secondary: #6f8565;
-$secondary: #000000;
-// $tertiary: #61876e;
-$tertiary: #19A7CE;
 
-$quad: #146C94;
+$primary: #a5243d;
+$secondary: #cee0dc;
+$tertiary: #b48291;
+$quad: #afaab9;
+$text: white;
 
 $box: 0 0 20px
   rgba(
